cleanupTasksAndSteps() moved to its own helper function
