+++ app/pages/lab-pages-editor/components/TasksPage/TasksPage.jsx
@@ -6,6 +6,7 @@ import getNewStepKey from '../../helpers/getNewStepKey.js';
 import getNewTaskKey from '../../helpers/getNewTaskKey.js';
 import linkStepsInWorkflow from '../../helpers/linkStepsInWorkflow.js';
 import moveItemInArray from '../../helpers/moveItemInArray.js';
+import cleanupTasksAndSteps from '../../helpers/cleanupTasksAndSteps.js';
 // import strings from '../../strings.json'; // TODO: move all text into strings
 
 import EditStepDialog from './components/EditStepDialog';
@@ -98,31 +99,6 @@ export default function TasksPage() {
 
     const cleanedTasksAndSteps = cleanupTasksAndSteps(newTasks, newSteps); 
     update(cleanedTasksAndSteps);
-  }
-
-  /*
-  Clean up tasks and steps.
-  - TODO: Remove steps without tasks.
-  - TODO: Remove tasks not associated with any step.
-  - Remove orphaned references in branching tasks.
-   */
-  function cleanupTasksAndSteps(tasks = {}, steps = []) {
-    const newTasks = structuredClone(tasks);  // Copy tasks
-    const newSteps = steps.slice();  // Copy steps
-
-    const taskKeys = Object.keys(newTasks);
-    const stepKeys = newSteps.map(step => step[0]);
-
-    Object.values(newTasks).forEach(taskBody => {
-      taskBody?.answers?.forEach(answer => {
-        // If the branching answer points to a non-existent Task Key or Step Key, remove the 'next'.
-        if (answer.next && !taskKeys.includes(answer.next) && !stepKeys.includes(answer.next)) {
-          delete answer.next;
-        }
-      })
-    });
-
-    return { tasks: newTasks, steps: newSteps };
   }
 
   // aka openEditStepDialog
